Moved INetworkedVar interface to NetworkedVarMeta file
