+++ MLAPI/Data/NetworkedVarMeta.cs
@@ -1,7 +1,67 @@
 ﻿using MLAPI.MonoBehaviours.Core;
+using MLAPI.NetworkingManagerComponents.Binary;
 
 namespace MLAPI.Data
 {
+
+    /// <summary>
+    /// Interface for networked value containers
+    /// </summary>
+    public interface INetworkedVar
+    {
+        /// <summary>
+        /// Returns the name of the channel to be used for syncing
+        /// </summary>
+        /// <returns>The name of the channel to be used for syncing</returns>
+        string GetChannel();
+        /// <summary>
+        /// Resets the dirty state and marks the variable as synced / clean
+        /// </summary>
+        void ResetDirty();
+        /// <summary>
+        /// Gets wheter or not the container is dirty
+        /// </summary>
+        /// <returns>Wheter or not the container is dirty</returns>
+        bool IsDirty();
+        /// <summary>
+        /// Gets wheter or not a specific client can write to the varaible
+        /// </summary>
+        /// <param name="clientId">The clientId of the remote client</param>
+        /// <returns>Wheter or not the client can write to the variable</returns>
+        bool CanClientWrite(uint clientId);
+        /// <summary>
+        /// Gets wheter or not a specific client can read to the varaible
+        /// </summary>
+        /// <param name="clientId">The clientId of the remote client</param>
+        /// <returns>Wheter or not the client can read to the variable</returns>
+        bool CanClientRead(uint clientId);
+        /// <summary>
+        /// Writes the dirty changes, that is, the changes since the variable was last dirty, to the writer
+        /// </summary>
+        /// <param name="writer">The writer to write the dirty changes to</param>
+        void WriteDelta(BitWriter writer);
+        /// <summary>
+        /// Writes the complete state of the variable to the writer
+        /// </summary>
+        /// <param name="writer">The writer to write the state to</param>
+        void WriteField(BitWriter writer);
+        /// <summary>
+        /// Reads the complete state from the reader and applies it
+        /// </summary>
+        /// <param name="reader">The reader to read the state from</param>
+        void ReadField(BitReader reader);
+        /// <summary>
+        /// Reads delta from the reader and applies them to the internal value
+        /// </summary>
+        /// <param name="reader">The reader to read the delta from</param>
+        void ReadDelta(BitReader reader);
+        /// <summary>
+        /// Sets NetworkedBehaviour the container belongs to.
+        /// </summary>
+        /// <param name="behaviour">The behaviour the container behaves to</param>
+        void SetNetworkedBehaviour(NetworkedBehaviour behaviour);
+    }
+
     /// <summary>
     /// The settings class used by the build in NetworkVar implementations
     /// </summary>
